Splitting code in modules
